Login관련 JS -> TS 로 마이그레이션
