+++ src/redux/loginSlice.ts
@@ -1,19 +1,27 @@
 import { createSlice } from "@reduxjs/toolkit";
 
 // API 서버 연동
-// reducer (store 상태 변경) 를 호출할때 지금은 API 호출
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { loginPost } from "../api/login/loginApi";
 import { getCookie, removeCookie, setCookie } from "../util/cookieUtil";
+import { LoginParam, LoginRes } from "../pages/admin/login/LoginPage";
 
-// export const 외부함수 = createAsyncThunk("이름", 리듀서함수);
 export const loginPostAsync = createAsyncThunk(
   "loginPostAsync",
-  async ({ loginParam, successFn, failFn, errorFn }) => {
+  async ({
+    loginParam,
+    successFn,
+    failFn,
+    errorFn,
+  }: {
+    loginParam: LoginParam;
+    successFn: (data: LoginRes) => void;
+    failFn: (error: string) => void;
+    errorFn: (error: string) => void;
+  }) => {
     try {
       const res = await loginPost({ loginParam, successFn, failFn, errorFn });
 
-      // 결과값을 리턴을 해야 action 에 값이 담기지...
       return res;
     } catch (error) {
       return error;
@@ -25,7 +33,6 @@ const initState = {
   nm: "",
 };
 
-// 쿠키 정보 읽어와서 initState 변경하기
 const loadMemberCookie = () => {
   const memberInfo = getCookie("nm");
   return memberInfo || initState;
